chore: move stateless transaction validation config to config file
