+++ crates/gateway/src/config.rs
@@ -40,3 +40,13 @@ impl Default for GatewayConfig {
         }
     }
 }
+
+#[derive(Clone, Default)]
+pub struct StatelessTransactionValidatorConfig {
+    // If true, validates that the resource bounds are not zero.
+    pub validate_non_zero_l1_gas_fee: bool,
+    pub validate_non_zero_l2_gas_fee: bool,
+
+    pub max_calldata_length: usize,
+    pub max_signature_length: usize,
+}
