+++ crates/gateway/src/gateway.rs
@@ -5,11 +5,9 @@ use starknet_api::external_transaction::ExternalTransaction;
 use std::net::SocketAddr;
 
 use crate::config::GatewayConfig;
-
+use crate::config::StatelessTransactionValidatorConfig;
 use crate::errors::GatewayError;
-use crate::stateless_transaction_validator::{
-    StatelessTransactionValidator, StatelessTransactionValidatorConfig,
-};
+use crate::stateless_transaction_validator::StatelessTransactionValidator;
 
 #[cfg(test)]
 #[path = "gateway_test.rs"]
