+++ crates/gateway/src/stateless_transaction_validator.rs
@@ -3,22 +3,13 @@ use starknet_api::external_transaction::{
 };
 use starknet_api::transaction::{Resource, ResourceBoundsMapping};
 
+use crate::config::StatelessTransactionValidatorConfig;
 use crate::errors::{StatelessTransactionValidatorError, StatelessTransactionValidatorResult};
 use crate::utils::ExternalTransactionExt;
 
 #[cfg(test)]
 #[path = "stateless_transaction_validator_test.rs"]
 mod stateless_transaction_validator_test;
-
-#[derive(Clone, Default)]
-pub struct StatelessTransactionValidatorConfig {
-    // If true, validates that the resource bounds are not zero.
-    pub validate_non_zero_l1_gas_fee: bool,
-    pub validate_non_zero_l2_gas_fee: bool,
-
-    pub max_calldata_length: usize,
-    pub max_signature_length: usize,
-}
 
 #[derive(Clone)]
 pub struct StatelessTransactionValidator {
