+++ crates/gateway/tests/routing_test.rs
@@ -2,8 +2,8 @@ use axum::body::{Body, Bytes, HttpBody};
 use axum::http::{Request, StatusCode};
 use pretty_assertions::assert_str_eq;
 use rstest::rstest;
+use starknet_gateway::config::StatelessTransactionValidatorConfig;
 use starknet_gateway::gateway::app;
-use starknet_gateway::stateless_transaction_validator::StatelessTransactionValidatorConfig;
 use std::fs;
 use std::path::Path;
 use tower::ServiceExt;
